requirements.txt
turbopuffer==0.1.13
pymongo==4.6.1
certifi==2023.11.17
numpy>=1.26.3,<2.0.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
voyageai==0.2.1
scikit-learn>=1.3.0